# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sheets/{id}/read:
    get:
      operationId: sheetRead
      tags: [sheets]
      summary: Read a Google Sheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sheet data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetReadResponse"
  /sheets/read-url:
    post:
      operationId: sheetReadByUrl
      tags: [sheets]
      summary: Read a shared Google Sheet by URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SheetReadByUrlRequest"
      responses:
        "200":
          description: Sheet data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetReadResponse"
  /sheets/{id}/write:
    post:
      operationId: sheetWrite
      tags: [sheets]
      summary: Write to a Google Sheet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SheetWriteRequest"
      responses:
        "200":
          description: Write result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetWriteResponse"
  /sheets/{id}:
    delete:
      operationId: sheetDelete
      tags: [sheets]
      summary: Delete a Google Sheet from Google Drive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
  /sheets/create-and-write:
    post:
      operationId: sheetCreateAndWrite
      tags: [sheets]
      summary: Create a new Google Sheet and write budget data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SheetCreateAndWriteRequest"
      responses:
        "200":
          description: Created spreadsheet info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SheetCreateAndWriteResponse"
...
components:
  schemas:
...
    Bill:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
          description: Monthly amount (negative = expense)
        dayOfMonth:
          type: ["integer", "null"]
          description: Day of month bill is due (null for weekly bills)
        category:
          type: string
          description: User-defined label for this bill, e.g. Rent, Phone Bill
        type:
          type: string
          enum: [balanced, fixed, weekly]
          description: How the bill is distributed across weeks
        color:
          type: string
          description: Color key for UI display, e.g. blue, green, orange
      required:
        - name
        - amount
        - category
        - type
    WeeklyBill:
      type: object
      properties:
        name:
          type: string
        amount:
          type: number
      required:
        - name
        - amount
    WeeklyBudget:
      type: object
      properties:
        weekLabel:
          type: string
          description: Human-readable label like "Budget from 3/5/26 to 3/11/26"
        startDate:
          type: string
          description: ISO date string for week start
        endDate:
          type: string
          description: ISO date string for week end
        openingBalance:
          type: number
          description: Amount remaining from previous week
        paycheck:
          type: number
          description: Paycheck received this week
        bills:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyBill"
        totalBills:
          type: number
          description: Sum of all bill line items for this week
        closingBalance:
          type: number
          description: Amount remaining after all bills
      required:
        - weekLabel
        - startDate
        - endDate
        - openingBalance
        - paycheck
        - bills
        - totalBills
        - closingBalance
...
    ExistingWeek:
      type: object
      properties:
        label:
          type: string
        startCol:
          type: number
        openingBalance:
          type: number
        paycheck:
          type: number
        items:
          type: array
          items:
            type: object
        remaining:
          type: number
      required:
        - label
        - startCol
        - openingBalance
        - paycheck
        - items
        - remaining
    SheetReadResponse:
      type: object
      properties:
        bills:
          type: array
          items:
            $ref: "#/components/schemas/Bill"
        existingWeeks:
          type: array
          items:
            $ref: "#/components/schemas/ExistingWeek"
        nextWeekStartCol:
          type: number
        lastRemaining:
          type: number
        sheetTitle:
          type: string
        spreadsheetId:
          type: string
      required:
        - bills
        - existingWeeks
        - nextWeekStartCol
        - lastRemaining
        - sheetTitle
    SheetReadByUrlRequest:
      type: object
      properties:
        url:
          type: string
      required:
        - url
    SheetWriteRequest:
      type: object
      properties:
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyBudget"
        billsSummary:
          type: array
          items:
            $ref: "#/components/schemas/Bill"
        sheetTitle:
          type: string
        startCol:
          type: number
        includeRemainingAcct:
          type: boolean
      required:
        - weeks
        - startCol
    SheetWriteResponse:
      type: object
      properties:
        success:
          type: boolean
        sheetTitle:
          type: string
      required:
        - success
    SheetCreateAndWriteRequest:
      type: object
      properties:
        title:
          type: string
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/WeeklyBudget"
        includeRemainingAcct:
          type: boolean
      required:
        - title
        - weeks
    SheetCreateAndWriteResponse:
      type: object
      properties:
        spreadsheetId:
          type: string
        spreadsheetUrl:
          type: string
      required:
        - spreadsheetId
        - spreadsheetUrl
...